Add files via upload
Student Database Management System Using Single Linked List

Implementation of student database management system using Single Linked List concept in Data Structure , color coding and makefile. Here, menu based program is implemented which asks user to select the option needed to perform
1) To Add New Record
2) Delete Record
3) Show List
4) Modify List
5) Save
6) Read from File
7) Exit
8) Sort List
9) Delete All Records 
10) Reverse List

To Add New Record:
There is an option to add name and marks of each student and roll number is updated automatically based on the input. Memory is allocated dynamically and link it appropriately into the singly linked list

Delete Record:
There are option to delete on the basic of name and rollno and also option to delete all links.

Show List:
Displaying the results in a tabular format

Modify List: 
There is an option to modify the list on the basic of name, roll number and percentage, If name/percentage is the input, there is a possibility of multiple existence, user should input the rollno which is associated with that name/percentage to modify.

Save & Read from File:
We can save the result in a file(Here, "student.dat" is the file) for further reference and also has the option to read from the file.
Exit:
Option to "save and exit"  and "exit without saving"

Sort List:
There are options to sort data based on name and percentage

Reverse  Link:
Using this option we are reversing the links